feat: KAN-300 Processor의 파싱 로직을 Factory로 분리
